Create LED Assembly Code
This code is in MIPS, machine code or assembly language, where it can be run on a plp tool. The code lights up the right 4 leds when switch 0 is set. It lights up the left 4 leds when switch 1 is set. Then when switch 2 is set, the code will loop through the leds lighting up only the odd numbers.